Created events routes and modified GETs to send API data.
