+++ src/server/app.js
@@ -1,12 +1,12 @@
 var app = require('express')();
 var bodyParser = require('body-parser');
 var github = require('./github-helper');
-var calendar = require('./calendar-helper');
+var googleHelper = require('./google-helper/');
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 github.initializeClubData().then((clubData) => {
-    calendar.initCalendar((calendar) => {
-        app.use('/', require('./routes')());
+    googleHelper.calendarHelper.initCalendar((calendar) => {
+        app.use('/', require('./routes')(calendar, clubData));
     });
 });
 app.listen(80);
