+++ src/server/routes/events/new.js
@@ -1,5 +1,5 @@
 module.exports = function (req, res) {
-    var googleCalendarHelper = require('../../calendar-helper');
+    var googleCalendarHelper = require('../../google-helper/').calendarHelper;
     googleCalendarHelper.insertEvent(req.body, (result) =>  {
         if (!result)    {
             res.sendStatus(401);
